+++ Camel.Engine/Tokens/NegationOperatorToken.cs
@@ -2,7 +2,7 @@ namespace Lepecki.Playground.Camel.Engine.Tokens
 {
     public class NegationOperatorToken : UnaryOperatorToken
     {
-        protected override double Calculate(double a)
+        protected override decimal Calculate(decimal a)
         {
             return -a;
         }
